Add output of traction change on fault to FaultCohesiveImpulses.
- Move traction change computation to FaultCohesive.
- Move auxiliary factory and derived factory to FaultCohesive.
