+++ libsrc/pylith/faults/FaultCohesive.cc
@@ -13,8 +13,11 @@
 #include "pylith/faults/FaultCohesive.hh" // implementation of object methods
 
 #include "pylith/faults/TopologyOps.hh" // USES TopologyOps
+#include "pylith/faults/AuxiliaryFieldFactory.hh" // USES AuxiliaryFieldFactory
 #include "pylith/faults/DiagnosticFieldFactory.hh" // USES DiagnosticFieldFactory
+#include "pylith/faults/DerivedFieldFactory.hh" // USES DerivedFieldFactory
 #include "pylith/fekernels/BoundaryDirections.hh" // USES BoundaryDirections
+#include "pylith/fekernels/FaultCohesiveKin.hh" // USES FaultCohesiveKin
 #include "pylith/feassemble/IntegratorInterface.hh" // USES IntegratorInterface
 #include "pylith/feassemble/InterfacePatches.hh" // USES InterfacePatches
 #include "pylith/feassemble/ConstraintSimple.hh" // USES ConstraintSimple
@@ -23,6 +26,7 @@
 #include "pylith/topology/Field.hh" // USES Field
 #include "pylith/topology/FieldOps.hh" // USES FieldOps::checkDiscretization()
 #include "pylith/topology/MeshOps.hh" // USES MeshOps::checkTopology()
+#include "pylith/topology/Stratum.hh" // USES Stratum
 
 #include "pylith/utils/error.hh" // USES PYLITH_METHOD_*
 #include "pylith/utils/journals.hh" // USES PYLITH_COMPONENT_*
@@ -39,7 +43,9 @@ typedef pylith::feassemble::IntegratorInterface::ProjectKernels ProjectKernels;
 // ------------------------------------------------------------------------------------------------
 // Default constructor.
 pylith::faults::FaultCohesive::FaultCohesive(void) :
+    _auxiliaryFactory(new pylith::faults::AuxiliaryFieldFactory),
     _diagnosticFactory(new pylith::faults::DiagnosticFieldFactory),
+    _derivedFactory(new pylith::faults::DerivedFieldFactory),
     _surfaceLabelName(""),
     _buriedEdgesLabelName(""),
     _surfaceLabelValue(1),
@@ -70,7 +76,9 @@ pylith::faults::FaultCohesive::deallocate(void) {
     PYLITH_METHOD_BEGIN;
     pylith::problems::Physics::deallocate();
 
+    delete _auxiliaryFactory;_auxiliaryFactory = NULL;
     delete _diagnosticFactory;_diagnosticFactory = NULL;
+    delete _derivedFactory;_derivedFactory = NULL;
 
     PYLITH_METHOD_END;
 } // deallocate
@@ -326,6 +334,56 @@ pylith::faults::FaultCohesive::createDiagnosticField(const pylith::topology::Fie
 
 
 // ------------------------------------------------------------------------------------------------
+// Create derived field.
+pylith::topology::Field*
+pylith::faults::FaultCohesive::createDerivedField(const pylith::topology::Field& solution,
+                                                  const pylith::topology::Mesh& domainMesh) {
+    PYLITH_METHOD_BEGIN;
+    PYLITH_COMPONENT_DEBUG("createDerivedField(solution="<<solution.getLabel()<<", domainMesh=)"<<typeid(domainMesh).name()<<")");
+
+    assert(_normalizer);
+
+    pylith::topology::Field* derivedField = new pylith::topology::Field(domainMesh);assert(derivedField);
+    derivedField->setLabel("FaultCohesive derived field");
+
+    // Create label for output.
+    const char* outputLabelName = "output";
+    PetscDM derivedDM = derivedField->getDM();
+    PetscDMLabel outputLabel = NULL;
+    PetscErrorCode err = PETSC_SUCCESS;
+    err = DMCreateLabel(derivedDM, outputLabelName);PYLITH_CHECK_ERROR(err);
+    err = DMGetLabel(derivedDM, outputLabelName, &outputLabel);PYLITH_CHECK_ERROR(err);
+    pylith::topology::Stratum faultStratum(derivedDM, pylith::topology::Stratum::HEIGHT, 1);
+    for (PetscInt point = faultStratum.begin(); point != faultStratum.end(); ++point) {
+        err = DMLabelSetValue(outputLabel, point, 1);
+    } // for
+    err = DMPlexLabelComplete(derivedDM, outputLabel);PYLITH_CHECK_ERROR(err);
+
+    assert(_derivedFactory);
+    const pylith::topology::FieldBase::Discretization& discretization = solution.getSubfieldInfo("lagrange_multiplier_fault").fe;
+    const PylithInt cellDim = solution.getSpaceDim()-1;
+    const bool isFaultOnly = false;
+    _derivedFactory->setSubfieldDiscretization("default", discretization.basisOrder, discretization.quadOrder, cellDim,
+                                               isFaultOnly, discretization.cellBasis, discretization.feSpace,
+                                               discretization.isBasisContinuous);
+
+    assert(_derivedFactory);
+    assert(_normalizer);
+    _derivedFactory->initialize(derivedField, *_normalizer, solution.getSpaceDim());
+
+    _derivedFactory->addTractionChange(); // 0
+
+    derivedField->subfieldsSetup();
+    derivedField->createDiscretization();
+    pylith::topology::FieldOps::checkDiscretization(solution, *derivedField);
+    derivedField->allocate();
+    derivedField->createOutputVector();
+
+    PYLITH_METHOD_RETURN(derivedField);
+} // createDerivedField
+
+
+// ------------------------------------------------------------------------------------------------
 // Create integrator and set kernels.
 pylith::feassemble::Integrator*
 pylith::faults::FaultCohesive::createIntegrator(const pylith::topology::Field& solution,
@@ -459,6 +517,22 @@ pylith::faults::FaultCohesive::_updateKernelConstants(const PylithReal dt) {
 
 
 // ------------------------------------------------------------------------------------------------
+// Get auxiliary factory associated with physics.
+pylith::feassemble::AuxiliaryFactory*
+pylith::faults::FaultCohesive::_getAuxiliaryFactory(void) {
+    return _auxiliaryFactory;
+} // _getAuxiliaryFactory
+
+
+// ------------------------------------------------------------------------------------------------
+// Get derived factory associated with fault.
+pylith::topology::FieldFactory*
+pylith::faults::FaultCohesive::_getDerivedFactory(void) {
+    return _derivedFactory;
+} // _getDerivedFactory
+
+
+// ------------------------------------------------------------------------------------------------
 pylith::feassemble::Integrator*
 pylith::faults::FaultCohesive::createIntegrator(const pylith::topology::Field& solution) {
     return NULL;
@@ -488,7 +562,7 @@ pylith::faults::FaultCohesive::_setKernelsDiagnosticField(pylith::feassemble::In
     integrator->setKernelsDiagnosticField(kernels);
 
     PYLITH_METHOD_END;
-}
+} // _setKernelsDiagnosticField
 
 
 // ------------------------------------------------------------------------------------------------
@@ -496,7 +570,20 @@ pylith::faults::FaultCohesive::_setKernelsDiagnosticField(pylith::feassemble::In
 void
 pylith::faults::FaultCohesive::_setKernelsDerivedField(pylith::feassemble::IntegratorInterface* integrator,
                                                        const pylith::topology::Field& solution) const {
-}
+    PYLITH_METHOD_BEGIN;
+    PYLITH_COMPONENT_DEBUG("_setKernelsDerivedField(integrator="<<integrator<<", solution="<<solution.getLabel()<<")");
+
+    const spatialdata::geocoords::CoordSys* coordsys = solution.getMesh().getCoordSys();
+    assert(coordsys);
+
+    std::vector<ProjectKernels> kernels(1);
+    kernels[0] = ProjectKernels("traction_change", pylith::fekernels::FaultCohesiveKin::tractionChange_asVector);
+
+    assert(integrator);
+    integrator->setKernelsDerivedField(kernels);
+
+    PYLITH_METHOD_END;
+} // _setKernelsDerivedField
 
 
 // End of file
