+++ libsrc/pylith/faults/FaultCohesiveKin.hh
@@ -58,16 +58,6 @@ public:
     pylith::topology::Field* createAuxiliaryField(const pylith::topology::Field& solution,
                                                   const pylith::topology::Mesh& domainMesh);
 
-    /** Create derived field.
-     *
-     * @param[in] solution Solution field.
-     * @param[in\ domainMesh Finite-element mesh associated with integration domain.
-     *
-     * @returns Derived field if applicable, otherwise NULL.
-     */
-    pylith::topology::Field* createDerivedField(const pylith::topology::Field& solution,
-                                                const pylith::topology::Mesh& domainMesh);
-
     /** Update auxiliary subfields at beginning of time step.
      *
      * @param[out] auxiliaryField Auxiliary field.
@@ -78,18 +68,6 @@ public:
 
     // PROTECTED METHODS //////////////////////////////////////////////////////////////////////////
 protected:
-
-    /** Get auxiliary factory associated with physics.
-     *
-     * @return Auxiliary factory for physics object.
-     */
-    pylith::feassemble::AuxiliaryFactory* _getAuxiliaryFactory(void);
-
-    /** Get derived factory associated with physics.
-     *
-     * @return Derived factory for physics object.
-     */
-    pylith::topology::FieldFactory* _getDerivedFactory(void);
 
     /** Update slip related subfields in auxiliary field at beginning of time step.
      *
@@ -137,8 +115,6 @@ protected:
     // PROTECTED MEMBERS //////////////////////////////////////////////////////////////////////////
 protected:
 
-    pylith::faults::AuxiliaryFieldFactory* _auxiliaryFactory; ///< Factory for auxiliary subfields.
-    pylith::faults::DerivedFieldFactory* _derivedFactory; ///< Factory for derived subfields.
     srcs_type _ruptures; ///< Array of kinematic earthquake ruptures.
     PetscVec _slipVecRupture; ///< PETSc local Vec to hold slip for one kinematic rupture.
     PetscVec _slipVecTotal; ///< PETSc local Vec to hold slip for all kinematic ruptures.
