+++ modulesrc/faults/FaultCohesiveKin.i
@@ -70,16 +70,6 @@ public:
             pylith::topology::Field* createAuxiliaryField(const pylith::topology::Field& solution,
                                                           const pylith::topology::Mesh& domainMesh);
 
-            /** Create derived field.
-             *
-             * @param[in] solution Solution field.
-             * @param[in\ domainMesh Finite-element mesh associated with integration domain.
-             *
-             * @returns Derived field if applicable, otherwise NULL.
-             */
-            pylith::topology::Field* createDerivedField(const pylith::topology::Field& solution,
-                                                        const pylith::topology::Mesh& domainMesh);
-
             /** Update auxiliary subfields at beginning of time step.
              *
              * @param[out] auxiliaryField Auxiliary field.
@@ -90,12 +80,6 @@ public:
 
             // PROTECTED METHODS //////////////////////////////////////////////////////////////////
 protected:
-
-            /** Get auxiliary factory associated with physics.
-             *
-             * @return Auxiliary factory for physics object.
-             */
-            pylith::feassemble::AuxiliaryFactory* _getAuxiliaryFactory(void);
 
             /** Update slip related subfields in auxiliary field at beginning of time step.
              *
